REF: misplaced repr tests
