+++ pandas/tests/io/formats/test_printing.py
@@ -1,11 +1,8 @@
+# Note! This file is aimed specifically at pandas.io.formats.printing utility
+# functions, not the general printing of pandas objects.
 import string
 
-import numpy as np
-import pytest
-
 import pandas._config.config as cf
-
-import pandas as pd
 
 from pandas.io.formats import printing
 
@@ -116,122 +113,3 @@ c        ff         いいい"""
         expected = "あ  dd    ggg \nb   ええ  ¡¡ab\nc   ff    いいい"
         adjoined = adj.adjoin(2, *data)
         assert adjoined == expected
-
-
-class TestTableSchemaRepr:
-    def test_publishes(self, ip):
-        ipython = ip.instance(config=ip.config)
-        df = pd.DataFrame({"A": [1, 2]})
-        objects = [df["A"], df]  # dataframe / series
-        expected_keys = [
-            {"text/plain", "application/vnd.dataresource+json"},
-            {"text/plain", "text/html", "application/vnd.dataresource+json"},
-        ]
-
-        opt = pd.option_context("display.html.table_schema", True)
-        last_obj = None
-        for obj, expected in zip(objects, expected_keys):
-            last_obj = obj
-            with opt:
-                formatted = ipython.display_formatter.format(obj)
-            assert set(formatted[0].keys()) == expected
-
-        with_latex = pd.option_context("styler.render.repr", "latex")
-
-        with opt, with_latex:
-            formatted = ipython.display_formatter.format(last_obj)
-
-        expected = {
-            "text/plain",
-            "text/html",
-            "text/latex",
-            "application/vnd.dataresource+json",
-        }
-        assert set(formatted[0].keys()) == expected
-
-    def test_publishes_not_implemented(self, ip):
-        # column MultiIndex
-        # GH 15996
-        midx = pd.MultiIndex.from_product([["A", "B"], ["a", "b", "c"]])
-        df = pd.DataFrame(
-            np.random.default_rng(2).standard_normal((5, len(midx))), columns=midx
-        )
-
-        opt = pd.option_context("display.html.table_schema", True)
-
-        with opt:
-            formatted = ip.instance(config=ip.config).display_formatter.format(df)
-
-        expected = {"text/plain", "text/html"}
-        assert set(formatted[0].keys()) == expected
-
-    def test_config_on(self):
-        df = pd.DataFrame({"A": [1, 2]})
-        with pd.option_context("display.html.table_schema", True):
-            result = df._repr_data_resource_()
-
-        assert result is not None
-
-    def test_config_default_off(self):
-        df = pd.DataFrame({"A": [1, 2]})
-        with pd.option_context("display.html.table_schema", False):
-            result = df._repr_data_resource_()
-
-        assert result is None
-
-    def test_enable_data_resource_formatter(self, ip):
-        # GH 10491
-        formatters = ip.instance(config=ip.config).display_formatter.formatters
-        mimetype = "application/vnd.dataresource+json"
-
-        with pd.option_context("display.html.table_schema", True):
-            assert "application/vnd.dataresource+json" in formatters
-            assert formatters[mimetype].enabled
-
-        # still there, just disabled
-        assert "application/vnd.dataresource+json" in formatters
-        assert not formatters[mimetype].enabled
-
-        # able to re-set
-        with pd.option_context("display.html.table_schema", True):
-            assert "application/vnd.dataresource+json" in formatters
-            assert formatters[mimetype].enabled
-            # smoke test that it works
-            ip.instance(config=ip.config).display_formatter.format(cf)
-
-
-# TODO: this is a test for Series/DataFrame __repr__
-@pytest.mark.parametrize(
-    "data,output",
-    [
-        ([2, complex("nan"), 1], [" 2.0+0.0j", " NaN+0.0j", " 1.0+0.0j"]),
-        ([2, complex("nan"), -1], [" 2.0+0.0j", " NaN+0.0j", "-1.0+0.0j"]),
-        ([-2, complex("nan"), -1], ["-2.0+0.0j", " NaN+0.0j", "-1.0+0.0j"]),
-        ([-1.23j, complex("nan"), -1], ["-0.00-1.23j", "  NaN+0.00j", "-1.00+0.00j"]),
-        ([1.23j, complex("nan"), 1.23], [" 0.00+1.23j", "  NaN+0.00j", " 1.23+0.00j"]),
-        (
-            [-1.23j, complex(np.nan, np.nan), 1],
-            ["-0.00-1.23j", "  NaN+ NaNj", " 1.00+0.00j"],
-        ),
-        (
-            [-1.23j, complex(1.2, np.nan), 1],
-            ["-0.00-1.23j", " 1.20+ NaNj", " 1.00+0.00j"],
-        ),
-        (
-            [-1.23j, complex(np.nan, -1.2), 1],
-            ["-0.00-1.23j", "  NaN-1.20j", " 1.00+0.00j"],
-        ),
-    ],
-)
-@pytest.mark.parametrize("as_frame", [True, False])
-def test_ser_df_with_complex_nans(data, output, as_frame):
-    # GH#53762, GH#53841
-    obj = pd.Series(np.array(data))
-    if as_frame:
-        obj = obj.to_frame(name="val")
-        reprs = [f"{i} {val}" for i, val in enumerate(output)]
-        expected = f"{'val': >{len(reprs[0])}}\n" + "\n".join(reprs)
-    else:
-        reprs = [f"{i}   {val}" for i, val in enumerate(output)]
-        expected = "\n".join(reprs) + "\ndtype: complex128"
-    assert str(obj) == expected, f"\n{str(obj)}\n\n{expected}"
